Removes storyboards

+++ Stex/ACAnswerCell.m
@@ -36,7 +36,7 @@
     if ([self.delegate respondsToSelector:@selector(userDidDownvoteAnswer:)])
         if (![self.delegate userDidDownvoteAnswer:self.postIDLabel.text])
         {
-            self.voteCountLabel.text = [NSString stringWithFormat:@"%d", voteCount];
+            self.voteCountLabel.text = [NSString stringWithFormat:@"%ld", (long)voteCount];
             [(UIButton *)sender setSelected:NO];
         }
     [(UIButton *)sender setSelected:YES];
@@ -59,7 +59,7 @@
     if ([self.delegate respondsToSelector:@selector(userDidUpvoteAnswer:)])
         if (![self.delegate userDidUpvoteAnswer:self.postIDLabel.text])
         {
-            self.voteCountLabel.text = [NSString stringWithFormat:@"%d", voteCount];
+            self.voteCountLabel.text = [NSString stringWithFormat:@"%ld", (long)voteCount];
             [(UIButton *)sender setSelected:NO];
         }
     [(UIButton *)sender setSelected:YES];
